restructure season pages
